Extremely rough first pass over the new edit role screen. Note that this new UI hasn't been ported to the new role screen yet.
This change introduces the ability to grant, deny, or remove individual capabilities for a role.  It also introduces a new method of grouping capabilities.

Several other key elements are the addition of a meta box system with custom meta boxes added, a tabbed interface for selecting caps, and new method for showing the role and role name.