ADDED actions at the top and bottom of the content permissions meatball.
This will allow us to hook in and modify what is shown before the roles list and docs. Contains WordPress standards compliant hook documentation. Potential use-case for this is for combining the members plugin with JJJ's User Groups plugin (meaning that we'd be able to limit content to a specific user group, rather than by a role)