This commit makes several changes dealing with editing roles on a multisite setup.  Primarily, the issue is that no user (even administrators) should have permission to edit roles in a multisite setup unless that user is a Super Admin.
The first step is that administrators are no longer granted the `create_roles`, `edit_roles`, and `delete_roles` capabilities when the plugin is activated.  Super admins can later grant these caps on a case-by-case basis if necessary.

Several network-specific capabilities are now hidden and cannot be added to any role, regardless of the user's permissions.

All hidden caps (including the newly-hidden network caps) are now blocked on both the PHP and JavaScript side of the edit/new role screen.

This change also removes the link to the edit role screen when a user doesn't have permission to actually edit the role (previously, it simply wouldn't load the screen).

Hat tip to @johnbillion.